pyproject.toml
[project]
name = "customers-data-updater"
version = "0.1.0"
description = "Multi-tenant SaaS for real estate data refresh via AI + WhatsApp"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "boto3>=1.35.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "openpyxl>=3.1.0",
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.5.0",
    "chardet>=5.2.0",
    "pandas>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[tool.mypy]
plugins = ["sqlalchemy.ext.mypy.plugin"]
ignore_missing_imports = true

[tool.ruff.lint.per-file-ignores]
"app/models/*.py" = ["F821"]  # SQLAlchemy forward-reference strings

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
